requirements.txt
appnope==0.1.4
asttokens==2.4.1
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
executing==2.1.0
Faker==33.0.0
ipykernel==6.29.5
ipython==8.29.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prettytable==3.12.0
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.5.2
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
